Add effects and filters screens to the block screen
